requirements.txt
joblib==1.3.2
numpy==1.26.4
scikit-learn==1.4.1.post1
scipy==1.12.0
threadpoolctl==3.3.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib